refactor: media view constructor
