+++ internal/domain/media/resource_media.go
@@ -24,13 +24,6 @@ type ResourceMedia struct {
 	DeletedAt    time.Time    `field:"deleted_at"`
 }
 
-type ResourceMediaView struct {
-	ID           int          `field:"id"`
-	ResourceType ResourceType `field:"resource_type"`
-	ResourceID   int          `field:"resource_id"`
-	MediaID      int          `field:"media_id"`
-}
-
 type ResourceMediaStore interface {
 	CreateResourceMedia(ctx context.Context, tx *database.Tx, resourceID int, mediaID int, resourceType string) (*ResourceMedia, *pnd.AppError)
 	FindResourceMediaByResourceID(ctx context.Context, tx *database.Tx, resourceID int, resourceType string) ([]Media, *pnd.AppError)
